+++ mex/common/models/__init__.py
@@ -1,11 +1,13 @@
-from typing import Final, Union, get_args
+from typing import Final, get_args
 
 from mex.common.models.access_platform import (
+    BaseAccessPlatform,
     ExtractedAccessPlatform,
     MergedAccessPlatform,
 )
 from mex.common.models.activity import (
     AdditiveActivity,
+    BaseActivity,
     ExtractedActivity,
     MergedActivity,
     PreventiveActivity,
@@ -13,10 +15,12 @@ from mex.common.models.activity import (
 )
 from mex.common.models.base import BaseModel
 from mex.common.models.contact_point import (
+    BaseContactPoint,
     ExtractedContactPoint,
     MergedContactPoint,
 )
 from mex.common.models.distribution import (
+    BaseDistribution,
     ExtractedDistribution,
     MergedDistribution,
 )
@@ -28,27 +32,32 @@ from mex.common.models.extracted_data import (
 )
 from mex.common.models.merged_item import MergedItem
 from mex.common.models.organization import (
+    BaseOrganization,
     ExtractedOrganization,
     MergedOrganization,
 )
 from mex.common.models.organizational_unit import (
+    BaseOrganizationalUnit,
     ExtractedOrganizationalUnit,
     MergedOrganizationalUnit,
 )
-from mex.common.models.person import ExtractedPerson, MergedPerson
+from mex.common.models.person import BasePerson, ExtractedPerson, MergedPerson
 from mex.common.models.primary_source import (
+    BasePrimarySource,
     ExtractedPrimarySource,
     MergedPrimarySource,
 )
-from mex.common.models.resource import ExtractedResource, MergedResource
+from mex.common.models.resource import BaseResource, ExtractedResource, MergedResource
 from mex.common.models.variable import (
     AdditiveVariable,
+    BaseVariable,
     ExtractedVariable,
     MergedVariable,
     PreventiveVariable,
     SubtractiveVariable,
 )
 from mex.common.models.variable_group import (
+    BaseVariableGroup,
     ExtractedVariableGroup,
     MergedVariableGroup,
 )
@@ -96,20 +105,37 @@ __all__ = (
     "MEX_PRIMARY_SOURCE_STABLE_TARGET_ID",
 )
 
+AnyBaseModel = (
+    BaseAccessPlatform
+    | BaseActivity
+    | BaseContactPoint
+    | BaseDistribution
+    | BaseOrganization
+    | BaseOrganizationalUnit
+    | BasePerson
+    | BasePrimarySource
+    | BaseResource
+    | BaseVariable
+    | BaseVariableGroup
+)
+BASE_MODEL_CLASSES: Final[list[type[AnyBaseModel]]] = list(get_args(AnyBaseModel))
+BASE_MODEL_CLASSES_BY_NAME: Final[dict[str, type[AnyBaseModel]]] = {
+    cls.__name__: cls for cls in BASE_MODEL_CLASSES
+}
 
-AnyExtractedModel = Union[
-    ExtractedAccessPlatform,
-    ExtractedActivity,
-    ExtractedContactPoint,
-    ExtractedDistribution,
-    ExtractedOrganization,
-    ExtractedOrganizationalUnit,
-    ExtractedPerson,
-    ExtractedPrimarySource,
-    ExtractedResource,
-    ExtractedVariable,
-    ExtractedVariableGroup,
-]
+AnyExtractedModel = (
+    ExtractedAccessPlatform
+    | ExtractedActivity
+    | ExtractedContactPoint
+    | ExtractedDistribution
+    | ExtractedOrganization
+    | ExtractedOrganizationalUnit
+    | ExtractedPerson
+    | ExtractedPrimarySource
+    | ExtractedResource
+    | ExtractedVariable
+    | ExtractedVariableGroup
+)
 EXTRACTED_MODEL_CLASSES: Final[list[type[AnyExtractedModel]]] = list(
     get_args(AnyExtractedModel)
 )
@@ -117,19 +143,19 @@ EXTRACTED_MODEL_CLASSES_BY_NAME: Final[dict[str, type[AnyExtractedModel]]] = {
     cls.__name__: cls for cls in EXTRACTED_MODEL_CLASSES
 }
 
-AnyMergedModel = Union[
-    MergedAccessPlatform,
-    MergedActivity,
-    MergedContactPoint,
-    MergedDistribution,
-    MergedOrganization,
-    MergedOrganizationalUnit,
-    MergedPerson,
-    MergedPrimarySource,
-    MergedResource,
-    MergedVariable,
-    MergedVariableGroup,
-]
+AnyMergedModel = (
+    MergedAccessPlatform
+    | MergedActivity
+    | MergedContactPoint
+    | MergedDistribution
+    | MergedOrganization
+    | MergedOrganizationalUnit
+    | MergedPerson
+    | MergedPrimarySource
+    | MergedResource
+    | MergedVariable
+    | MergedVariableGroup
+)
 MERGED_MODEL_CLASSES: Final[list[type[AnyMergedModel]]] = list(get_args(AnyMergedModel))
 MERGED_MODEL_CLASSES_BY_NAME: Final[dict[str, type[AnyMergedModel]]] = {
     cls.__name__: cls for cls in MERGED_MODEL_CLASSES