+++ mex/common/models/activity.py
@@ -27,7 +27,8 @@ from mex.common.types import (
     MergedPrimarySourceIdentifier,
     Text,
     Theme,
-    Timestamp,
+    YearMonth,
+    YearMonthDay,
 )
 
 
@@ -41,9 +42,7 @@ class _OptionalLists(BaseModel):
     ] = []
     alternativeTitle: list[Text] = []
     documentation: list[Link] = []
-    end: list[
-        Annotated[Timestamp, Field(examples=["2024-01-17", "2024", "2024-01"])]
-    ] = []
+    end: list[YearMonthDay | YearMonth] = []
     externalAssociate: list[MergedOrganizationIdentifier | MergedPersonIdentifier] = []
     funderOrCommissioner: list[MergedOrganizationIdentifier] = []
     fundingProgram: list[str] = []
@@ -52,9 +51,7 @@ class _OptionalLists(BaseModel):
     isPartOfActivity: list[MergedActivityIdentifier] = []
     publication: list[Link] = []
     shortName: list[Text] = []
-    start: list[
-        Annotated[Timestamp, Field(examples=["2023-01-16", "2023", "2023-02"])]
-    ] = []
+    start: list[YearMonthDay | YearMonth] = []
     succeeds: list[MergedActivityIdentifier] = []
     theme: list[
         Annotated[Theme, Field(examples=["https://mex.rki.de/item/theme-1"])]