+++ mex/common/types/__init__.py
@@ -1,4 +1,4 @@
-from typing import Final, Union, get_args
+from typing import Final, get_args
 
 from mex.common.types.email import Email
 from mex.common.types.identifier import (
@@ -30,15 +30,17 @@ from mex.common.types.identity import IdentityProvider
 from mex.common.types.link import Link, LinkLanguage
 from mex.common.types.path import AssetsPath, PathWrapper, WorkPath
 from mex.common.types.sink import Sink
-from mex.common.types.text import Text, TextLanguage
-from mex.common.types.timestamp import (
+from mex.common.types.temporal_entity import (
     CET,
-    TIMESTAMP_FORMATS_BY_PRECISION,
-    TIMESTAMP_REGEX,
+    TEMPORAL_ENTITY_FORMATS_BY_PRECISION,
     UTC,
-    Timestamp,
-    TimestampPrecision,
+    TemporalEntity,
+    TemporalEntityPrecision,
+    YearMonth,
+    YearMonthDay,
+    YearMonthDayTime,
 )
+from mex.common.types.text import Text, TextLanguage
 from mex.common.types.vocabulary import (
     AccessRestriction,
     ActivityType,
@@ -109,38 +111,37 @@ __all__ = (
     "Text",
     "TextLanguage",
     "Theme",
-    "TIMESTAMP_FORMATS_BY_PRECISION",
-    "TIMESTAMP_REGEX",
-    "Timestamp",
-    "TimestampPrecision",
+    "TEMPORAL_ENTITY_FORMATS_BY_PRECISION",
+    "TemporalEntity",
+    "YearMonth",
+    "YearMonthDay",
+    "YearMonthDayTime",
+    "TemporalEntityPrecision",
     "UTC",
     "VocabularyEnum",
     "VocabularyLoader",
     "WorkPath",
 )
 
-AnyNestedModel = Union[
-    Link,
-    Text,
-]
+AnyNestedModel = Link | Text
 NESTED_MODEL_CLASSES: Final[list[type[AnyNestedModel]]] = list(get_args(AnyNestedModel))
 NESTED_MODEL_CLASSES_BY_NAME: Final[dict[str, type[AnyNestedModel]]] = {
     cls.__name__: cls for cls in NESTED_MODEL_CLASSES
 }
 
-AnyMergedIdentifier = Union[
-    MergedAccessPlatformIdentifier,
-    MergedActivityIdentifier,
-    MergedContactPointIdentifier,
-    MergedDistributionIdentifier,
-    MergedOrganizationalUnitIdentifier,
-    MergedOrganizationIdentifier,
-    MergedPersonIdentifier,
-    MergedPrimarySourceIdentifier,
-    MergedResourceIdentifier,
-    MergedVariableGroupIdentifier,
-    MergedVariableIdentifier,
-]
+AnyMergedIdentifier = (
+    MergedAccessPlatformIdentifier
+    | MergedActivityIdentifier
+    | MergedContactPointIdentifier
+    | MergedDistributionIdentifier
+    | MergedOrganizationalUnitIdentifier
+    | MergedOrganizationIdentifier
+    | MergedPersonIdentifier
+    | MergedPrimarySourceIdentifier
+    | MergedResourceIdentifier
+    | MergedVariableGroupIdentifier
+    | MergedVariableIdentifier
+)
 MERGED_IDENTIFIER_CLASSES: Final[list[type[AnyMergedIdentifier]]] = list(
     get_args(AnyMergedIdentifier)
 )
@@ -148,19 +149,19 @@ MERGED_IDENTIFIER_CLASSES_BY_NAME: Final[dict[str, type[AnyMergedIdentifier]]] =
     cls.__name__: cls for cls in MERGED_IDENTIFIER_CLASSES
 }
 
-AnyExtractedIdentifier = Union[
-    ExtractedAccessPlatformIdentifier,
-    ExtractedActivityIdentifier,
-    ExtractedContactPointIdentifier,
-    ExtractedDistributionIdentifier,
-    ExtractedOrganizationalUnitIdentifier,
-    ExtractedOrganizationIdentifier,
-    ExtractedPersonIdentifier,
-    ExtractedPrimarySourceIdentifier,
-    ExtractedResourceIdentifier,
-    ExtractedVariableGroupIdentifier,
-    ExtractedVariableIdentifier,
-]
+AnyExtractedIdentifier = (
+    ExtractedAccessPlatformIdentifier
+    | ExtractedActivityIdentifier
+    | ExtractedContactPointIdentifier
+    | ExtractedDistributionIdentifier
+    | ExtractedOrganizationalUnitIdentifier
+    | ExtractedOrganizationIdentifier
+    | ExtractedPersonIdentifier
+    | ExtractedPrimarySourceIdentifier
+    | ExtractedResourceIdentifier
+    | ExtractedVariableGroupIdentifier
+    | ExtractedVariableIdentifier
+)
 EXTRACTED_IDENTIFIER_CLASSES: Final[list[type[AnyExtractedIdentifier]]] = list(
     get_args(AnyExtractedIdentifier)
 )