+++ CHANGELOG.md
@@ -19,6 +19,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.22.0] - 2024-03-19
+
+### Added
+
+- port `get_inner_types` from `mex-backend` to `mex.common.utils`
+
+### Changes
+
+- rename Timestamp class to TemporalEntity
+- added subclasses with specific resolution YearMonth, YearMonthDay, YearMonthDayTime
+- modernize typing with syntactic sugar
+- simplify `BaseModel._get_list_field_names` using `get_inner_types`
+- switch from poetry to pdm
+- use vocabulary JSON files from mex-model
+
+### Removed
+
+- remove vocabulary JSON files
+
+### Fixed
+
+- date and time validation working and harmonized with mex-model
+
 ## [0.21.0] - 2024-03-04
 
 ### Added